+++ lib/utils/genId.ts
@@ -7,3 +7,25 @@ const genId = (len = 20) => {
     return id;
 };
 export { genId };
+
+if (import.meta.vitest) {
+    const { describe, expect, it } = import.meta.vitest;
+
+    describe("genId", () => {
+        it("should generate a random id", () => {
+            const id = genId();
+            expect(id).toBeTruthy();
+        });
+
+        it("should generate a random id with a specific length", () => {
+            const id = genId(10);
+            expect(id).toHaveLength(10);
+        });
+
+        it("should truly generate a random id several times", () => {
+            const ids = Array.from({ length: 100 }, () => genId());
+            expect(new Set(ids).size).toBe(ids.length);
+        })
+    });
+
+}
