+++ tsconfig.json
@@ -2,6 +2,9 @@
   "compilerOptions": {
     "target": "ES2020",
     "useDefineForClassFields": true,
+    "types": [
+      "vitest/importMeta"
+    ],
     "lib": [
       "ES2020",
       "DOM",
